• MappingProjection   - refactoring to replace function with Identity Function for IDENTITY_MATRIX     works.
