+++ psyneulink/core/components/projections/modulatory/learningprojection.py
@@ -167,9 +167,8 @@ import typecheck as tc
 
 from psyneulink.core.components.component import parameter_keywords
 from psyneulink.core.components.functions.function import is_function_type
-from psyneulink.core.components.functions.interfacefunctions import Identity
 from psyneulink.core.components.functions.learningfunctions import BackPropagation
-from psyneulink.core.components.functions.transferfunctions import Linear
+from psyneulink.core.components.functions.transferfunctions import Linear, Identity
 from psyneulink.core.components.functions.combinationfunctions import LinearCombination
 from psyneulink.core.components.mechanisms.adaptive.learning.learningmechanism import ERROR_SIGNAL, LearningMechanism
 from psyneulink.core.components.mechanisms.processing.objectivemechanism import ObjectiveMechanism
