+++ tests/functions/test_identity.py
@@ -1,4 +1,5 @@
 import numpy as np
+import psyneulink.core.components.functions.transferfunctions
 import pytest
 
 import psyneulink.core.components.functions.interfacefunctions as Functions
@@ -11,7 +12,7 @@ import psyneulink.core.llvm as pnlvm
 @pytest.mark.benchmark(group="IdentityFunction")
 def test_basic(size, benchmark):
     variable = np.random.rand(size)
-    f = Functions.Identity(default_variable=variable)
+    f = psyneulink.core.components.functions.transferfunctions.Identity(default_variable=variable)
     res = benchmark(f.function, variable)
     assert np.allclose(res, variable)
 
@@ -22,7 +23,7 @@ def test_basic(size, benchmark):
 @pytest.mark.benchmark(group="IdentityFunction")
 def test_llvm(size, benchmark):
     variable = np.random.rand(size)
-    f = Functions.Identity(default_variable=variable)
+    f = psyneulink.core.components.functions.transferfunctions.Identity(default_variable=variable)
     m = pnlvm.execution.FuncExecution(f)
     res = benchmark(m.execute, variable)
     assert np.allclose(res, variable)
@@ -35,7 +36,7 @@ def test_llvm(size, benchmark):
 @pytest.mark.benchmark(group="IdentityFunction")
 def test_ptx_cuda(size, benchmark):
     variable = np.random.rand(size)
-    f = Functions.Identity(default_variable=variable)
+    f = psyneulink.core.components.functions.transferfunctions.Identity(default_variable=variable)
     m = pnlvm.execution.FuncExecution(f)
     res = benchmark(m.cuda_execute, variable)
     assert np.allclose(res, variable)
